all the auth is done (i thiiiiiiink?)

+++ Submission/src/DLObject.java
@@ -98,12 +98,47 @@ public class DLObject {
     public ArrayList<ArrayList<String>> fetch(String tableName, String pkName, String pkData, Jws<Claims> token) throws DLException{
         MySQLDatabase db = new MySQLDatabase(uName, uPass);
         if (db.connect()) {
-            String sql = "SELECT * FROM " + tableName + " WHERE " + pkName + " = ?;";
-            ArrayList<String> values = new ArrayList<>();
-            values.add(pkData);
-            return db.getData(sql, values);
+            int isAdmin = Integer.parseInt((String) token.getBody().get("IsAdmin"));
+            int loginUserId = Integer.parseInt((String) token.getBody().get("UserID"));
+
+            boolean hasAccess = false;
+
+            if ((!tableName.equals("Papers")) && (!tableName.equals("Users")) && isAdmin != 1) {
+                System.out.println("Access Denied: Only Admins can do that!");
+            } else if (isAdmin == 1) {
+                hasAccess = true;
+            } else {
+                if (tableName.equals("Users")) {
+                    if (loginUserId == Integer.parseInt(pkData)) {
+                        hasAccess = true;
+                    }
+                } else {
+                    String query = "SELECT * FROM PaperAuthors WHERE paperId = ?;";
+                    ArrayList<String> paperIdHolder = new ArrayList<>();
+                    paperIdHolder.add(pkData);
+                    ArrayList<ArrayList<String>> fullResults = db.getData(query, paperIdHolder);
+                    ArrayList<ArrayList<String>> results = new ArrayList<>();
+                    for (int i = 2; i < fullResults.size(); i++) {
+                        results.add(fullResults.get(i));
+                    }
+                    for (int i = 0; i < results.size(); i++) {
+                        if (loginUserId == Integer.parseInt(results.get(i).get(0))) {
+                            hasAccess = true;
+                        }
+                    }
+                }
+            }
+            if (hasAccess) {
+                String sql = "SELECT * FROM " + tableName + " WHERE " + pkName + " = ?;";
+                ArrayList<String> values = new ArrayList<>();
+                values.add(pkData);
+                return db.getData(sql, values);
+            } else {
+                return new ArrayList<>();
+            }
         }
         else {
+            System.out.println("Failed to connect!");
             return new ArrayList<>();
         }
     }
@@ -119,23 +154,58 @@ public class DLObject {
      */
     public int put(String tableName, ArrayList<String> columnNames, ArrayList<String> values, int numKeys, Jws<Claims> token) throws DLException{
         MySQLDatabase db = new MySQLDatabase(uName, uPass);
-        if (db.connect()){
-            StringBuilder sql = new StringBuilder("UPDATE " + tableName + " SET ");
-            for (int i = 0; i < columnNames.size(); i++){
-                String column = columnNames.get(i);
-                if (i == (columnNames.size() - numKeys)){
-                    sql.append("WHERE ").append(column).append(" = ?;");
-                }
-                else if (i == ((columnNames.size() - numKeys) - 1)){
-                    sql.append(column).append(" = ? ");
-                }
-                else {
-                    sql.append(column).append(" = ?, ");
+        if (db.connect()) {
+            int isAdmin = Integer.parseInt((String) token.getBody().get("IsAdmin"));
+            int loginUserId = Integer.parseInt((String) token.getBody().get("UserID"));
+
+            boolean hasAccess = false;
+
+            if ((!tableName.equals("Papers")) && (!tableName.equals("Users")) && isAdmin != 1) {
+                System.out.println("Access Denied: Only Admins can do that!");
+            } else if (isAdmin == 1) {
+                hasAccess = true;
+            } else {
+                if (tableName.equals("Users")) {
+                    if (loginUserId == Integer.parseInt(values.get(values.size() - 1))) {
+                        hasAccess = true;
+                    }
+                } else {
+                    String query = "SELECT * FROM PaperAuthors WHERE paperId = ?;";
+                    ArrayList<String> paperIdHolder = new ArrayList<>();
+                    paperIdHolder.add(values.get(values.size() - 1));
+
+                    ArrayList<ArrayList<String>> fullResults = db.getData(query, paperIdHolder);
+                    ArrayList<ArrayList<String>> results = new ArrayList<>();
+
+                    for (int i = 2; i < fullResults.size(); i++) {
+                        results.add(fullResults.get(i));
+                    }
+                    for (int i = 0; i < results.size(); i++) {
+                        if (loginUserId == Integer.parseInt(results.get(i).get(0))) {
+                            hasAccess = true;
+                        }
+                    }
                 }
             }
-            return db.setData(sql.toString(), values);
+            if (hasAccess) {
+                StringBuilder sql = new StringBuilder("UPDATE " + tableName + " SET ");
+                for (int i = 0; i < columnNames.size(); i++) {
+                    String column = columnNames.get(i);
+                    if (i == (columnNames.size() - numKeys)) {
+                        sql.append("WHERE ").append(column).append(" = ?;");
+                    } else if (i == ((columnNames.size() - numKeys) - 1)) {
+                        sql.append(column).append(" = ? ");
+                    } else {
+                        sql.append(column).append(" = ?, ");
+                    }
+                }
+                return db.setData(sql.toString(), values);
+            } else {
+                return -1;
+            }
         }
         else {
+            System.out.println("Failed to connect!");
             return -1;
         }
     }
@@ -152,18 +222,55 @@ public class DLObject {
     public int post(String tableName, ArrayList<String> columnNames, ArrayList<String> values, Jws<Claims> token) throws DLException{
         MySQLDatabase db = new MySQLDatabase(uName, uPass);
         if (db.connect()) {
-            StringBuilder sql = new StringBuilder("INSERT INTO " + tableName + " VALUES ( ");
-            for (int i = 0; i < columnNames.size(); i++) {
-                String column = columnNames.get(i);
-                if (i == columnNames.size() - 1) {
-                    sql.append(column).append("?);");
+            int isAdmin = Integer.parseInt((String) token.getBody().get("IsAdmin"));
+            int loginUserId = Integer.parseInt((String) token.getBody().get("UserID"));
+
+            boolean hasAccess = false;
+
+            if ((!tableName.equals("Papers")) && (!tableName.equals("Users")) && isAdmin != 1) {
+                System.out.println("Access Denied: Only Admins can do that!");
+            } else if (isAdmin == 1) {
+                hasAccess = true;
+            } else {
+                if (tableName.equals("Users")) {
+                    if (loginUserId == Integer.parseInt(values.get(0))) {
+                        hasAccess = true;
+                    }
                 } else {
-                    sql.append(column).append("?, ");
+                    String query = "SELECT * FROM PaperAuthors WHERE paperId = ?;";
+                    ArrayList<String> paperIdHolder = new ArrayList<>();
+                    paperIdHolder.add(values.get(0));
+
+                    ArrayList<ArrayList<String>> fullResults = db.getData(query, paperIdHolder);
+                    ArrayList<ArrayList<String>> results = new ArrayList<>();
+
+                    for (int i = 2; i < fullResults.size(); i++) {
+                        results.add(fullResults.get(i));
+                    }
+                    for (int i = 0; i < results.size(); i++) {
+                        if (loginUserId == Integer.parseInt(results.get(i).get(0))) {
+                            hasAccess = true;
+                        }
+                    }
                 }
             }
-            return db.setData(sql.toString(), values);
+            if (hasAccess) {
+                StringBuilder sql = new StringBuilder("INSERT INTO " + tableName + " VALUES ( ");
+                for (int i = 0; i < columnNames.size(); i++) {
+                    String column = columnNames.get(i);
+                    if (i == columnNames.size() - 1) {
+                        sql.append(column).append("?);");
+                    } else {
+                        sql.append(column).append("?, ");
+                    }
+                }
+                return db.setData(sql.toString(), values);
+            } else {
+                return -1;
+            }
         }
         else {
+            System.out.println("Failed to connect!");
             return -1;
         }
     }
@@ -180,20 +287,54 @@ public class DLObject {
     public int delete(String tableName, ArrayList<String> pkNames, ArrayList<String> pkData, Jws<Claims> token) throws DLException{
         MySQLDatabase db = new MySQLDatabase(uName, uPass);
         if (db.connect()) {
-            StringBuilder sql = new StringBuilder("DELETE FROM " + tableName + " WHERE ");
-            for (int i = 0; i < pkNames.size(); i++){
-                String pk = pkNames.get(i);
-                if (i == pkNames.size() - 1){
-                    sql.append(pk).append(" = ?;");
-                }
-                else {
-                    sql.append(pk).append(" = ? AND ");
+            int isAdmin = Integer.parseInt((String) token.getBody().get("IsAdmin"));
+            int loginUserId = Integer.parseInt((String) token.getBody().get("UserID"));
+
+            boolean hasAccess = false;
+
+            if ((!tableName.equals("Papers")) && (!tableName.equals("Users")) && isAdmin != 1) {
+                System.out.println("Access Denied: Only Admins can do that!");
+            } else if (isAdmin == 1) {
+                hasAccess = true;
+            } else {
+                if (tableName.equals("Users")) {
+                    if (loginUserId == Integer.parseInt(pkData.get(0))) {
+                        hasAccess = true;
+                    }
+                } else {
+                    String query = "SELECT * FROM PaperAuthors WHERE paperId = ?;";
+                    ArrayList<String> paperIdHolder = new ArrayList<>();
+                    paperIdHolder.add(pkData.get(0));
+                    ArrayList<ArrayList<String>> fullResults = db.getData(query, paperIdHolder);
+                    ArrayList<ArrayList<String>> results = new ArrayList<>();
+                    for (int i = 2; i < fullResults.size(); i++) {
+                        results.add(fullResults.get(i));
+                    }
+                    for (int i = 0; i < results.size(); i++) {
+                        if (loginUserId == Integer.parseInt(results.get(i).get(0))) {
+                            hasAccess = true;
+                        }
+                    }
                 }
             }
+            if (hasAccess) {
+                StringBuilder sql = new StringBuilder("DELETE FROM " + tableName + " WHERE ");
+                for (int i = 0; i < pkNames.size(); i++) {
+                    String pk = pkNames.get(i);
+                    if (i == pkNames.size() - 1) {
+                        sql.append(pk).append(" = ?;");
+                    } else {
+                        sql.append(pk).append(" = ? AND ");
+                    }
+                }
 
-            return db.setData(sql.toString(), pkData);
+                return db.setData(sql.toString(), pkData);
+            } else {
+                return -1;
+            }
         }
         else {
+            System.out.println("Failed to connect!");
             return -1;
         }
     }
@@ -210,12 +351,47 @@ public class DLObject {
     public int delete(String tableName, String pkName, String pkData, Jws<Claims> token) throws DLException{
         MySQLDatabase db = new MySQLDatabase(uName, uPass);
         if (db.connect()) {
-            String sql = "DELETE FROM " + tableName + " WHERE " + pkName + " = ?;";
-            ArrayList<String> values = new ArrayList<>();
-            values.add(pkData);
-            return db.setData(sql, values);
+            int isAdmin = Integer.parseInt((String) token.getBody().get("IsAdmin"));
+            int loginUserId = Integer.parseInt((String) token.getBody().get("UserID"));
+
+            boolean hasAccess = false;
+
+            if ((!tableName.equals("Papers")) && (!tableName.equals("Users")) && isAdmin != 1) {
+                System.out.println("Access Denied: Only Admins can do that!");
+            } else if (isAdmin == 1) {
+                hasAccess = true;
+            } else {
+                if (tableName.equals("Users")) {
+                    if (loginUserId == Integer.parseInt(pkData)) {
+                        hasAccess = true;
+                    }
+                } else {
+                    String query = "SELECT * FROM PaperAuthors WHERE paperId = ?;";
+                    ArrayList<String> paperIdHolder = new ArrayList<>();
+                    paperIdHolder.add(pkData);
+                    ArrayList<ArrayList<String>> fullResults = db.getData(query, paperIdHolder);
+                    ArrayList<ArrayList<String>> results = new ArrayList<>();
+                    for (int i = 2; i < fullResults.size(); i++) {
+                        results.add(fullResults.get(i));
+                    }
+                    for (int i = 0; i < results.size(); i++) {
+                        if (loginUserId == Integer.parseInt(results.get(i).get(0))) {
+                            hasAccess = true;
+                        }
+                    }
+                }
+            }
+            if (hasAccess) {
+                String sql = "DELETE FROM " + tableName + " WHERE " + pkName + " = ?;";
+                ArrayList<String> values = new ArrayList<>();
+                values.add(pkData);
+                return db.setData(sql, values);
+            } else {
+                return -1;
+            }
         }
         else {
+            System.out.println("Failed to connect!");
             return -1;
         }
     }
